Initialize project as a Python project.
This was initialized using `python-boilerplate`.  It should give you all of the
boilerplate code you need for a basic Python package that could be put in PyPI.
It'll be a while before we're there and I really just wanted the boilerplate for
a Python module, not full-fledged package.  But I'll try this.